Create palindrome_number.py
First, we convert the integer x to a string using the str function. Then, we check if the string is equal to its reverse by using slicing with a step of -1 to reverse the string. Finally, we return True if the string is a palindrome, and False otherwise.

Note that we use the -> bool type hint to indicate that the function returns a boolean value. We also use the int type hint to indicate that the argument x is an integer.